devicetree: Fix Device tree tests cannot be built with coverage enabled
To Peproduce:
./scripts/twister -p native_sim -T tests/lib/devicetree --coverage-tool gcovr -i --coverage

Logs:
INFO - Using Ninja..
INFO - Zephyr version: v3.7.0-1236-gad3e941ad3ed
INFO - Using 'zephyr' toolchain.
INFO - Building initial testsuite list...
INFO - Writing JSON report xxx/zephyr/twister-out/testplan.json
INFO - JOBS: 22
INFO - Adding tasks to the queue...
INFO - Added initial list of jobs to queue
ERROR - native_sim tests/lib/devicetree/api/libraries.devicetree.api ERROR : Build failure
INFO - xxx/zephyr/twister-out/native_sim/tests/lib/devicetree/api/libraries.devicetree.api/build.log
ERROR - Loading Zephyr default modules (Zephyr base).
/usr/bin/ld: xxx/zephyr/tests/lib/devicetree/api/src/main.c:1190: undefined reference to __device_dts_ord_86'
/usr/bin/ld: xxx/zephyr/tests/lib/devicetree/api/src/main.c:1197: undefined reference to `__device_dts_ord_87'
collect2: error: ld returned 1 exit status

Reason:
DT_SPEC related macro tests (ADC_DT_SPEC_GET_BY_NAME, MBOX_DT_SPEC_GET)should not appear in this test because the configuration does not turn on any devices

Move DP_SPEC-related tests to api_ext and enable the related device in the configuration

fix #77205
