+++ tests/lib/devicetree/api_ext/prj.conf
@@ -1,1 +1,3 @@
 CONFIG_ZTEST=y
+CONFIG_ADC=y
+CONFIG_MBOX=y
